blip

+++ Makefile
@@ -70,7 +70,7 @@ install-base: sanity-check ## Install base packages
 	# Clean up, Update, Basics
 	sudo sed -e "s#.*ParallelDownloads.*#ParallelDownloads = 10#g" -i /etc/pacman.conf
 	echo 'BUILDDIR="/dev/shm/makepkg"' | sudo tee /etc/makepkg.conf.d/00-skillarch.conf
-	sudo cachyos-rate-mirrors # Increase install speed & Update repos
+	[ ! -f /.dockerenv ] && sudo cachyos-rate-mirrors || true # Increase install speed & Update repos (skip in Docker)
 	yes|sudo pacman -Scc
 	yes|sudo pacman -Syu
 	$(PACMAN_INSTALL) git vim tmux wget curl archlinux-keyring
